# backend/swagger.yaml
openapi: 3.0.0
info:
  title: User API
  version: 1.0.0

components:
  securitySchemes:
    bearerAuth:            # שם הסכימה
      type: http
      scheme: bearer
      bearerFormat: JWT    # ציון סוג הטוקן, אם מדובר ב-JWT

paths:
  /users/signUp:
    post:
      summary: Add a new user
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userName:
                  type: string
                userPassword:
                  type: string
                userEmail:
                  type: string
              required:
                - userName
                - userPassword
                - userEmail
      responses:
        '201':
          description: Successfully added user
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                  email:
                    type: string
                  token:
                    type: string

  /users/login:
    post:
      summary: User login
      tags:
        - Users
      description: Authenticates a user and returns a token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                loginEmail:
                  type: string
                  format: email
                  example: "user@example.com"
                loginPassword:
                  type: string
                  format: password
                  example: "userpassword"
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                      userName:
                        type: string
                      userEmail:
                        type: string
                        format: email
                  token:
                    type: string
                    example: "your.jwt.token.here"
  /business/getBusiness:
    get:
      summary: Get all businesses
      tags:
        - Business
      description: Retrieves a list of all businesses in the system
      security:
        - bearerAuth: [] 
      responses:
        '200':
          description: Successfully retrieved the list of businesses
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      example: 1
                    businessName:
                      type: string
                      example: "My Business"
                    businessDescription:
                      type: string
                      example: "This is a description of my business."
                    businessEmail:
                      type: string
                      example: "business@example.com"
                    businessPhone:
                      type: string
                      example: "+1234567890"
        '500':
          description: Internal Server Error
          content:
            text/plain:
              example: "Internal Server Error"
        '400':
          description: Bad Request
          content:
            text/plain:
              example: "Invalid query parameters"

  /business/addBusiness:
    post:
      summary: Add a new business
      tags:
        - Business
      security:
        - bearerAuth: []
      description: Adds a new business to the system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                businessName:
                  type: string
                  example: "My Business"
                businessDescription:
                  type: string
                  example: "This is a description of my business."
                businessEmail:
                  type: string
                  example: "business@example.com"
                businessPhone:
                  type: string
                  example: "+1234567890"
      responses:
        '201':
          description: Successfully added
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 2
                  businessName:
                    type: string
                    example: "My Business"
                  businessDescription:
                    type: string
                    example: "This is a description of my business."
                  businessEmail:
                    type: string
                    example: "business@example.com"
                  businessPhone:
                    type: string
                    example: "+1234567890"

  /business/updateBusiness/{id}:
    put:
      summary: Update an existing business
      tags:
        - Business
      security:
        - bearerAuth: []
      description: Updates an existing business in the system
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                businessName:
                  type: string
                  example: "Updated Business"
                businessDescription:
                  type: string
                  example: "This is an updated description of my business."
                businessEmail:
                  type: string
                  example: "updated_business@example.com"
                businessPhone:
                  type: string
                  example: "+0987654321"
      responses:
        '200':
          description: Successfully updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 2
                  businessName:
                    type: string
                    example: "Updated Business"
                  businessDescription:
                    type: string
                    example: "This is an updated description of my business."
                  businessEmail:
                    type: string
                    example: "updated_business@example.com"
                  businessPhone:
                    type: string
                    example: "+0987654321"

  /services/addService:
    post:
      summary: Add a new service
      tags:
        - Services
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Service'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'

  /services/updateService/{serviceId}:
    put:
      summary: Update an existing service
      tags:
        - Services
      security:
        - bearerAuth: []
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:  
            schema:
              $ref: '#/components/schemas/Service'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'

  /services/getServices:
    get:
      summary: Get all services
      tags:
        - Services
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Service'

  /services/deleteServiceById/{serviceId}:
    delete:
      summary: Delete a service by ID
      tags:
        - Services
      security:
        - bearerAuth: []
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /meetings/getMeetings:
    get:
      summary: Get all meetings
      tags:
        - Meetings
      security:
        - bearerAuth: []
      description: Retrieve a list of all meetings in the system.
      responses:
        '200':
          description: A list of meetings was successfully retrieved.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Meeting'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Internal Server Error

  /meetings/addMeeting:
    post:
      summary: Add a new meeting
      tags:
        - Meetings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Meeting'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Meeting'
  /meetings/updateMeeting/{meetingId}:
    put:
      summary: Update an existing meeting
      tags:
        - Meetings
      security:
        - bearerAuth: []
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:  
            schema:
              $ref: '#/components/schemas/Meeting'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Meeting'

  /meetings/deleteMeetingById/{MeetingId}:
    delete:
      summary: Delete a meeting by ID
      tags:
        - Meetings
      security:
        - bearerAuth: []
      parameters:
        - name: MeetingId
          in: path
          required: true
          description: The ID of the meeting to delete
          schema:
            type: integer
            example: 123
      responses:
        '200':
          description: Meeting deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Meeting deleted successfully"
        '400':
          description: Invalid request or missing parameters
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid Meeting ID"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Internal Server Error"

components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        userName:
          type: string
        userEmail:
          type: string
        userPassword:
          type: string
        userToken:
          type: string
      required:
        - userName
        - userEmail
        - userPassword

    Business:
      type: object
      properties:
        id:
          type: integer
        businessName:
          type: string
        businessDescription:
          type: string
        businessEmail:
          type: string
        businessPhone:
          type: string
      required:
        - businessName
        - businessDescription
        - businessEmail
        - businessPhone

    Service:
      type: object
      properties:
        id:
          type: integer
        serviceName:
          type: string
        serviceDescription:
          type: string
        serviceCost:
          type: number
        serviceDuration:
          type: integer
      required:
        - serviceName
        - serviceDescription
        - serviceCost
        - serviceDuration

    Meeting:
      type: object
      properties:
        id:
          type: integer
        meetingDate:
          type: string
          format: date-time
        userId:
          type: integer
        serviceId:
          type: integer
      required:
        - meetingDate
        - userId
        - serviceId